+++ run-jest-tests.sh
@@ -1,0 +1,6 @@
+for f in ./src/**/**/*.test.ts; do
+  NODE_OPTIONS=--experimental-vm-modules npx jest $f || exit 1;
+done
+for f in ./src/**/*.test.ts; do
+  NODE_OPTIONS=--experimental-vm-modules npx jest $f || exit 1;
+done
