+++ src/mina-signer/src/Utils.ts
